+++ ShoppingApplication/DTO/Product/CreateProductDTO.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 namespace Application.DTO.Product
 {
@@ -10,6 +7,6 @@ namespace Application.DTO.Product
         public string Name { get; set; }
         public string Description { get; set; }
         public Guid CategoryId { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
